libaio ***no_ci***

+++ fio.lab/libaio.lab/try-all.sh
@@ -1,6 +1,7 @@
 # work=/transient-builds/fio; sudo mkdir -p $work; sudo chown -R $(whoami) $work; cd $work; git clone https://github.com/devizer/KernelManagementLab; cd KernelManagementLab; git pull; cd fio.lab/libaio.lab; bash try-all.sh
 cat ver-links.txt | while read -r ver link; do
-    echo "$ver" "$link"
+    if [[ -z "$ver" ]]; then continue; fi
+    echo "libaio $ver: $link"
     work=bin/$ver
     mkdir -p $work
     rm -rf $work/*
